entity: move immutable data logic into trait
